Clean up card mapping
Agent-Logs-Url: https://github.com/vitorhugo-java/the_universe_decides/sessions/89ab9a3c-d319-472d-aad8-9f85b029f429

+++ lib/controllers/card_draw_controller.dart
@@ -33,7 +33,7 @@ class PlayingCard {
   });
 
   factory PlayingCard.fromNumber(int value) {
-    final normalized = value.clamp(1, 52) as int;
+    final normalized = value.clamp(1, 52);
     final zeroBased = normalized - 1;
 
     return PlayingCard(
